setupenvLnx.sh: Don't hard code TI GT specific paths
We should not be hard coding TI GT specific paths into the source files.
We should not make the user edit a source file to do thier build.
This change allows the user to predefine C6X_BASE_DIR.
If not defined it will look in standard places for CCS as is done for Windows.

Signed-off-by: Bill Mills <wmills@ti.com>